filesystem refactoring and removal of cruft
