+++ src/clients/vector/__init__.py
@@ -1,4 +1,4 @@
-from .company_report_client import CompanyReportClient
+from .vector_company_report_client import CompanyReportClient
 from .vector_report_client import VectorReportClient
 from .concept_decomposer import ConceptDecomposer
