+++ src/clients/vector/vector_company_report_client.py
