+++ src/constants/patterns/__init__.py
@@ -1,9 +1,8 @@
 """
 Pattern constants
 """
+
 __all__ = [
-    "BIOLOGIC_INFIXES",
-    "BIOLOGIC_SUFFIXES",
     "INDICATION_REGEXES",
     "INDICATION_MODIFIER_REGEXES",
     "IUPAC_RE",
@@ -11,18 +10,12 @@ __all__ = [
     "MOA_INFIXES",
     "MOA_PREFIXES",
     "MOA_SUFFIXES",
-    "PHASE_RE",
-    "SMALL_MOLECULE_INFIXES",
-    "SMALL_MOLECULE_SUFFIXES",
 ]
 
-from .biologics import BIOLOGIC_INFIXES, BIOLOGIC_SUFFIXES
 from .indication import INDICATION_MODIFIER_REGEXES, INDICATION_REGEXES
 from .moa import (
     MOA_INFIXES,
     MOA_PREFIXES,
     MOA_SUFFIXES,
 )
-from .other import PHASE_RE
-from .small_molecules import SMALL_MOLECULE_INFIXES, SMALL_MOLECULE_SUFFIXES
 from .iupac import IUPAC_RE, IUPAC_STRINGS
