--- src/constants/patterns/small_molecules.py
@@ -1,40 +1,0 @@
-SMALL_MOLECULE_SUFFIXES = {
-    "aban": "e.g. apixaban",
-    "amine": "amine compound",
-    "axel": "e.g. paclitaxel",
-    # "ide": "compound or drug",
-    # "ium": "cation or compound",
-    # "ol": "alcohol or compound", # FP: control
-    # "ate": "ester or compound",
-    "ium": "metal or compound",
-    "azole": "azole antifungal",
-    # "ine": "alkaloid or compound", # azacitidine
-    "dine": "e.g. azacitidine",
-    # "one": "ketone or compound", # FP: cornerstone
-    # "ide": "compound or drug",
-    # "in": "alkaloid or compound",
-    "ium": "metal or compound",
-    "mib": "Small molecules that work inside cancer cells to slow proliferation and increase apoptosis",
-    "oxin": "compound or drug",
-    "nib": "small-molecule inhibitor",  # e.g. lorlatinib maleate
-    "rin": "compound or drug",
-    "mten": "e.g. mavacamten",
-    "xian": "e.g. milvexian",
-    "oran": "e.g. afimetoran",
-    "xate": "e.g. methotrexate",
-    # "ten": "e.g. mavacamten" # FP: often
-    # Add more suffixes as needed
-}
-
-SMALL_MOLECULE_INFIXES = {
-    # "anti": "antibiotic or compound",
-    "barb": "barbiturate or compound",
-    "fluoro": "fluorinated compound",
-    "statin": "cholesterol-lowering drug",
-    "corti": "corticosteroid or compound",
-    "cyclo": "cyclic compound",
-    "thio": "thiol or compound",
-    "benz": "benzene ring or compound",
-    "phyto": "phytochemical or compound",
-    # Add more infixes as needed
-}
