+++ src/data/etl/entity/owner/test/test_owner_utils.py
@@ -2,8 +2,8 @@ import unittest
 
 from prisma.enums import OwnerType
 from pydash import group_by
-from constants.company import COMPANY_MAP
 
+from constants.company import COMPANY_MAP
 from data.etl.entity.owner.owner import generate_clean_owner_map, OwnerTypeParser
