+++ src/nlp/classifier/__init__.py
@@ -1,0 +1,3 @@
+from .classifier import classify_by_keywords, classify_string, create_lookup_map
+
+__all__ = ["classify_by_keywords", "classify_string", "create_lookup_map"]
