+++ src/nlp/clustering/__init__.py
@@ -1,0 +1,4 @@
+from .clustering import cluster_terms
+from .topic import TopicModeler
+
+__all__ = ["cluster_terms", "TopicModeler"]
