+++ src/nlp/nel/__init__.py
@@ -1,11 +1,11 @@
-from .candidate_selector.composite_candidate_selector import (
-    CompositeCandidateSelector,
-)
+from .candidate_selector.composite_candidate_selector import CompositeCandidateSelector
+from .candidate_selector.types import CandidateSelectorType
 from .candidate_selector.candidate_selector import CandidateSelector
 from .linker import TermLinker
 
 __all__ = [
     "CandidateSelector",
     "CompositeCandidateSelector",
+    "CandidateSelectorType",
     "TermLinker",
 ]
