+++ src/nlp/ner/__init__.py
@@ -1,13 +1,12 @@
 """
 NER module
 """
+
 # from .classifier import classify_by_keywords
 from .ner import NerTagger
-from .linker.linker import TermLinker
 from .normalizer import TermNormalizer
 
 __all__ = [
     "NerTagger",
-    "TermLinker",
     "TermNormalizer",
 ]
