--- src/nlp/topic/__init__.py
@@ -1,3 +1,0 @@
-from .topic import Topics
-
-__all__ = ["Topics"]
